week05-advanced-rag/minseon/requirements.txt
openai
chromadb==0.4.24
langchain-text-splitters
pymupdf
python-dotenv
streamlit
rank-bm25
